add config File to repo
