+++ just-paai-frontend-code/.gitignore
@@ -37,6 +37,3 @@ next-env.d.ts
 
 #env
 .env
-
-# config
-/config
